+++ packages/expo-router/src/native-tabs/index.ts
@@ -1,2 +1,11 @@
-export { Label, Icon, Badge } from './common/elements';
+export * from './common/elements';
+export type {
+  NativeTabsProps,
+  NativeTabTriggerProps,
+  NativeTabsStyleType,
+  NativeTabsActiveStyleType,
+  NativeTabOptions,
+  SfSymbolOrImageSource,
+} from './NativeBottomTabs/types';
+export { NativeTabTrigger } from './NativeBottomTabs/NativeTabTrigger';
 export { NativeTabs } from './NativeBottomTabs/NativeTabs';